agregando metodos crud simples en book-service

+++ book-service/src/main/java/com/library/bookservice/repository/BookRepository.java
@@ -3,7 +3,10 @@ package com.library.bookservice.repository;
 import com.library.bookservice.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
 
-public interface BookRepository extends JpaRepository<Book,Integer> {
+
+public interface BookRepository extends JpaRepository<Book,Long> {
+
 
 }
